Add functional test for searching a regulation
This commit adds a browser test that checks the functionality of searching a regulation. Specifically, the tests check to make sure that clicking the search icon in the nav panel displays the search bar, searching a term displays a results page, and clicking a search result takes the user to the appropriate subsection.