Fix functional test for search in IE 10
The functional test for search runs in Sauce Labs in a window that is too small to accommodate the 1366x768 resolution that was being used in this test. Also, the IE window cut off the result link that the test was trying to click, making it unclickable and making the test error.

Both problems are now fixed: the window is set to 1024x600 (as in the table of contents test) and the last result link on the first page of results—which is reliably always in the viewport and clickable—is now being tested.