use self-hosted fonts from cf.gov
This commit deprecates use of the font files in the private "eregsip"
repository, instead using font files located at /static/fonts that are
provided by cf.gov. Additionally, the ability to add these fonts locally
is provided via a new static.in directory, akin to how cfgov-refresh
does it. See documentation here:

https://cfpb.github.io/cfgov-refresh/installation/#webfonts

I've updated the README to make a note of this.