toggle reg update messages via settings
This commit adds support for a new Django setting,
EREGS_REGULATION_UPDATES, that allows for placement of a message on
certain regulations to let users know that an update in progress. This
allows for a central place to toggle these updates instead of requiring
changes on each reg's specific landing page.

Implementation is done via a new {% update_in_progress %} template tag
that can be passed a regulation part, that checks the hardcoded list in
settings.

This change also removes the hardcoded update message that currently
appears on part 1026 (Reg Z) and so any deployments will need to add the
setting like this to maintain current behavior:

\# in Django settings file
EREGS_REGULATION_UPDATES = ['1026']