remove use of threads when retrieving layers
This change removes the use of multiple threads when retrieving layer
JSON. Instead of spawning a new thread for each of the ten layers that
need to be retrieved on a request, the layers are retrieved
sequentially.

While this may have the effect of slowing down data retrieval somewhat,
it greatly simplifies how predictably this code can be expected to run
in a web server.