use nvm explicitly to enforce npm version
This change modifies how the Node version is specified in the Travis CI
build. The "nodejs: 6" entry in .travis.yml wasn't properly setting the
Node version to the desired 6.x version. This was causing problems
because Travis was using Node 8.x, which modified the
npm-shrinkwrap.json file. This in turn created a "dirty" git state which
caused release wheels to be named improperly.

Ping @higs4281 who I know was experiencing something similar in another
repository.

This repository should probably be upgraded to explicitly use Node 8
instead of Node 6, but that is left for a future effort.