Use label_id in reg_updates template tag
To show an update-in-progress message on any page of a regulation, we need to be able to use the reg_updates template tag in a few different templates. Context only contains reg_part in a few specific templates, so we'll change reg_updates to use label_id instead, which is available everywhere.