Add update-in-progress notifications to reg text
The notification that we're in the progress of adding notices to a regulation currently only appears on the landing page of affected regulation. That's a problem if you (a) don't notice or don't remember the notification on the landing page or (b) don't come to the regulation through the landing page (e.g. from a search).

To make it clearer everywhere that updates are in progress, this commit adds the update-in-progress notification to the sidebar of all types of pages in a regulation: reg text, interpretations, appendices, diffs, section-by-section analysis, and search results.